None: Fix implementation of `kDIRECT_IO`.
The previous implementation allowed falling back to I/O reformatting
if a reformat-free solution did not exist. That was contrary to
the design document and public API comments.

Make the implementation obey the specification.
Add public API comments on the difference betweek `kDIRECT_IO`
and the related legacy implication of `kSTRICT_TYPES`.

For samples/common/sampleEngines.cpp, use `kDIRECT_IO`
for DLA only if the user does not want fallback to the GPU.