http://nvbugs/3394293: Fix header definitions for fMHA
There is no explicit handshake for the fMHA kernels, so when the new
kernels were added from 1ead0fa8, the header file wasn't updated
leading to a mismatch on the host parameter size and the device
parameter size.

Also fixed some CC violations, used unsigned modulus instead of signed
modulus, and fixed the default initialization of the structure so that
memset isn't required.