None: Update Python docs for add_reduce, add_top_k, and ISoftMaxLayer.
The documentation for the axes parameter was years out of date,
and did not reflect changes with the introduction of "full dims".

I copied the text from `include/NvInfer.h` and adjusted the formatting
markup, and replaced `reduceAxes` with `axes` where appropriate.

Thanks to Ian Tramble for alerting me to the add_reduce documentation.