https://nvbugs/3518812 - TRT crashes on Windows with OSS C&R plugin
The MultilevelCropAndResize plugin was causing a crash only on Windows.
On further investigation, it was revealed that the crash was due to the
user's model only providing 4 FP inputs to the plugin instead of the
required 5 FPs as specified by the plugin. The reason for the crash to
only occur in Windows and not in Linux is still not well understood.

Although there is an assertion to check whether the correct number of
inputs are provided, it is a C-assertion, which does not get hit
in release builds. It was verified that the assertion is indeed hit in
the debug build. As such, this issue is expected to be preventable after the
completion of TRT-15919 (https://jirasw.nvidia.com/browse/TRT-15919), when
PLUGIN_ASSERT will replace C-assertions.

As a tentative resolution, the README of the MultilevelCropAndResize plugin
has been updated to highlight that the plugin expects 5 FPs. Missing
documentation for the `image_size` attribute has also been added.

Signed-off-by: skarunaratne <skarunaratne@nvidia.com>