1583 - sublicense ieee/half.h under Apache2
Apache2 is a more restrictive license than MIT, and requires that the
attributions to original author to be preserved.

MIT license allows sublicensing the software under a different license
provided that the copyright and permission notices are retained. As such,
software licensed under MIT can be digested into TensorRT-OSS project
(licensed under Apache2) with proper attributions to the original author.

Signed-off-by: Rajeev Rao <rajeevrao@nvidia.com>