https://nvbugs/3641975 - Support Windows paths through expliclity quoting input names
Add logic for supporting `'inputName':PATH` input format for trtexec's commandline options.

In trtexec documentation we state that input names can be wrapped in single quotes '', but we don't actually support that.
On Windows, --loadInputs flag will generally take a path that starts with `c:\..`, so splitting on last seen colon does not work in this case.