https://nvbugs/3682549 - BatchedNMSDynamicPlugin illegal memory access
Fixes a CUDA illegal memory access in the BatchedNMSDynamicPlugin.

The issue was an incorrect calculation in getWorkspaceSize() which
lead to less memory being allocated for the plugin than the plugin
expected. Below is the chain of events which caused the error:

1. configurePlugin() is called before inference can begin. It is given
an input tensor size with some dimensions being dynamic (size == -1).
It uses these dimensions to calculate member variables boxesSize,
scoresSize, and numPriors.

2. getWorkspaceSize() is called, and it uses boxesSize, scoresSize,
and numPriors for its calculation. However, these variables have
meaningless values because they were calculated with some dimensions
being -1.

3. TensorRT allocates the (incorrect) desired memory for the plugin.

The fixed version works as follows:

1. configurePlugin() is called with some dimensions being -1.
It uses these dimensions to calculate member variables boxesSize,
scoresSize, and numPriors, as before

2. getWorkspaceSize() is called, but it does not use boxesSize,
scoresSize, and numPriors for its calculation. Instead, it calculates
these values using the dimensions of the given PluginTensorDesc,
which are set to a value greater or equal to the (not yet known)
real input dimensions.

3. TensorRT allocates enough memory for the plugin.

A few more things:
- Remove duplicate sentence in README
- Rename member variables (e.g. scoresSize -> mScoresSize)
- In createPlugin, use the new validateRequiredAttributesExist function.