TRT-18235: create a shared library that contains ROIAlign plugin
For version forward compatibility, we want to generate a shared library which contains vfc plugins. Currently we only have ROIAlign plugin in the list.

This is a combination of the following changes:
- Generate libnvinfer_vfc_plugin.so file which conains ROIAlign plugin and is guarded by NV_BUILD_LPM=1.
- Move kernels.cpp/.h to common/kernels and exclude common/kernels/ from bc_plugin for reducing the .so size from 3.7M to 836K.
- Remove unused dependencies in building bc_plugin: lcublas, lcudnn, and lmyelin.
- Changes to remove the linking against lnvinfer (getLogger) and remove lnvinfer from build.
- Fix all west-const issues in kernels.h and change int to int32_t in kernels.cpp.
- Add tests for verifying the shared library is generated with LPM.

The size can be reduced to ~100K if cudart is linked dynamically, but cannot do this in production.

Size of shared libraries:
- release.86/libnvinfer_plugin.so.8.6.0: 14M
- release.86/libnvinfer_vfc_plugin.so.8.6.0: 836K