Run trt_mod_test subtests in parallel
Although the trt_mod_test framework allowed for subtests
to execute in parallel, this was not happening because
the tests were being run behind a lock.

This MR fixes this by running the subtest before acquiring
the lock, which is only used to record the results.

Also fix a minor bug in plugin attribute validation where
the error string was built incorrectly.