TRTINFRA-4096 - add CUDA12.0 builds to L* pipelines
This MR fixes a few API changes in CUDA to support CUDA12.0, and adds CUDA12.0 builds to L0/L1/LR.

The two API changes are that the `saturate` function was removed and `cudaGraphInstantiate` was changed to match the parameters of `cudaGraphInstantiateWithFlags`.
We already had an implementation for `saturate` that compiles correctly, although I'm not sure if it behaves the same way.

This MR also makes the CUDA version checks in the pwgen build system more robust.