TRT-18406: add APIs for loading plugins from buffers or lib files
This is a combination of the following commits:
- 5f3f0026: Add APIs for loading plugins from buffers or lib files

This MR addes the APIs in pluginRegistry for loading plugins from buffers or lib files. Main changes:

Add new APIs in class IPluginRegistry
- PluginLibraryHandle loadLibrary(char const* pluginPath);
- PluginLibraryHandle deserializeLibrary(void const* pluginBuffer, size_t size);
- void deregisterLibrary(PluginLibraryHandle handle);
- Add the corresponding implementations in standard and safety.

Add new test  in testing/runtime/rt/pluginRegistryTests.cpp
- Load the libnvinfer_vfc_plugin.so from disk.
- Check whether the shared library contains the ROIAlign plugin.
- Use the API to register the library and check whether the ROIAlign plugin can be found in the registry.
- Use the API to deregister the library and check whether ROIAlign plugin is deregistered.

Add extern “C” to the library entry points for loading the library symbol.