Memory issues in PriorBoxPlugin
This commit fixes a memory error in the PriorBoxPlugin
where make_unique<float>(size) was being used to allocate
an array of length 'size', when in reality it allocates
a single float and assigns it the value 'size'. Because of
the nature of memory allocation and 'size' usually being
small, this usually went unnoticed.

This commit also cleans up the code in a few different ways:
1. int -> int32_t
2. char -> uint8_t (when used as a byte in a buffer)
3. west const -> east const
4. Remove 'new' and 'delete' where possible.
5. PLUGIN_ASSERT -> PLUGIN_VALIDATE (no abort()-ing).