Fix bug when reading yaml file
The test_polygraphy_golden_ref[CustomGeluPluginDynamic_PluginConfig.yaml]
test failed on a QA machine because the Gelu PluginConfig
yaml file used UTF-8 encoding instead of Ascii, and python
tried to read it as ascii when we called open().

This commit fixes this in two ways:

1. Convert CustomGeluPluginDynamic_PluginConfig.yaml to ascii encoding.

2. Change the call to open() to read the yaml file in binary mode.
The pyyaml module can figure out the encoding on its own.